Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -66,7 +66,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view("questions.edit", compact('question'));
+      if(\Gate::denies('update-question', $question)){
+
+        abort(403,"Access denied");
+      }
+      return view("questions.edit", compact('question'));
     }
 
     /**
@@ -78,6 +82,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+
+        abort(403,"Access denied");
+        }
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', "Your question has been updated.");
@@ -91,8 +99,12 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+      if(\Gate::denies('delete-question', $question)){
+
+        abort(403,"Access denied");
+      }
         $question->delete();
 
-        return redirect('/questions')->with('success','Your question has been deleted');
+        return redirect('/questions')->with('success',"Your question has been deleted");
     }
 }
